requirements.txt
Django>=5.0
djangorestframework
psycopg[binary]
python-dotenv
djangorestframework-simplejwt
openpyxl
reportlab
django-filter
python-dateutil